+++ sources/tests/http-1.1.ts
@@ -40,7 +40,7 @@ router.use(async function(context, next) {
 
         try {
 
-            console.log("xxx");
+            console.log(`Requesting user ${context.params.id}`);
             await next();
             console.log("responsed");
 
@@ -75,7 +75,11 @@ router.use(async function(context, next) {
 
     context.response.write(`Requested at ${new Date(context.request.time)}`);
 
-}).get("/users/{user:int}", async function(context) {
+}).get("/users/{id:hex-string[5]}", async function(context) {
+
+    context.response.sendJSON(context.params);
+
+}).get(new RegExp("^/article/(.+)$"), async function(context) {
 
     context.response.sendJSON(context.params);